tweak ts config
