+++ package.json
@@ -135,6 +135,7 @@
     "@stryker-mutator/core": "6.3.0",
     "@stryker-mutator/jest-runner": "6.3.0",
     "@types/jest": "29.2.2",
+    "@types/sinon": "^10.0.13",
     "@types/tap": "15.0.7",
     "commitizen": "4.2.5",
     "cz-emoji": "1.3.1",
