+++ packages/diff/tsconfig.json
@@ -1,8 +1,23 @@
 {
   "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist",
+    "rootDir": ".",
+    "outDir": "./dist"
   },
-  "include": ["./src"]
+  "include": ["./src/**/*.ts"],
+  "ts-node": {
+    "files": true,
+    "esm": true,
+    "experimentalSpecifierResolution": "node",
+    "compilerOptions": {
+      "module": "ES2022"
+    },
+    "moduleTypes": {
+      "src/**/*": "esm",
+      "tests/**/*": "esm"
+    }
+  },
+  "references": [
+    { "path": "./packages/core" }
+  ]
 }
